requirements.txt
paho-mqtt
smbus
lockfile
bme680
